DIMboss_updateState: 99.20->99.34 (#14 (int) cast on masked status-flags test -> cmpwi not cmplwi)

+++ src/main/dll/DIM/dll_01E0_dimboss.c
@@ -340,7 +340,7 @@ int DIMboss_updateState(DIMbossObject* obj, undefined4 param_2, ObjAnimUpdateSta
             mapDirIndex = mapGetDirIdx(DIMTOP_MAP_DIR);
             mapLoadDataFile(mapDirIndex, DIMTOP_AUDIO_DATA_FILE_B);
             loadWaitStarted = false;
-            while (statusFlags = getLoadedFileFlags(0), (statusFlags & DIMTOP_LOAD_PENDING_FLAGS_MASK) != 0)
+            while (statusFlags = getLoadedFileFlags(0), (int)(statusFlags & DIMTOP_LOAD_PENDING_FLAGS_MASK) != 0)
             {
                 padUpdate();
                 checkReset();
